+++ src/systems/joint_controller/JointController.cc
@@ -1,0 +1,145 @@
+/*
+ * Copyright (C) 2019 Open Source Robotics Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+#include <ignition/msgs/double.pb.h>
+#include <ignition/plugin/Register.hh>
+#include <ignition/transport/Node.hh>
+
+#include "ignition/gazebo/components/JointVelocity.hh"
+#include "ignition/gazebo/Model.hh"
+
+#include "JointController.hh"
+
+using namespace ignition;
+using namespace gazebo;
+using namespace systems;
+
+class ignition::gazebo::systems::JointControllerPrivate
+{
+  /// \brief Callback for velocity subscription
+  /// \param[in] _msg Velocity message
+  public: void OnCmdVel(const ignition::msgs::Double &_msg);
+
+  /// \brief Ignition communication node.
+  public: transport::Node node;
+
+  /// \brief Joint Entity
+  public: Entity jointEntity;
+
+  /// \brief Joint name
+  public: std::string jointName;
+
+  /// \brief Commanded joint velocity
+  public: double jointVelCmd;
+
+  /// \brief mutex to protect jointVelCmd
+  public: std::mutex jointVelCmdMutex;
+
+  /// \brief Model interface
+  public: Model model{kNullEntity};
+};
+
+//////////////////////////////////////////////////
+JointController::JointController()
+  : dataPtr(std::make_unique<JointControllerPrivate>())
+{
+}
+
+//////////////////////////////////////////////////
+void JointController::Configure(const Entity &_entity,
+    const std::shared_ptr<const sdf::Element> &_sdf,
+    EntityComponentManager &_ecm,
+    EventManager &/*_eventMgr*/)
+{
+  this->dataPtr->model = Model(_entity);
+
+  if (!this->dataPtr->model.Valid(_ecm))
+  {
+    ignerr << "JointController plugin should be attached to a model entity. "
+           << "Failed to initialize." << std::endl;
+    return;
+  }
+
+  // Get params from SDF
+  this->dataPtr->jointName = _sdf->Get<std::string>("joint_name");
+
+  if (this->dataPtr->jointName == "")
+  {
+    ignerr << "JointController found an empty jointName parameter. "
+           << "Failed to initialize.";
+    return;
+  }
+
+  // Subscribe to commands
+  std::string topic{"/model/" + this->dataPtr->model.Name(_ecm) + "/joint/" +
+                    this->dataPtr->jointName + "/cmd_vel"};
+  this->dataPtr->node.Subscribe(topic, &JointControllerPrivate::OnCmdVel,
+                                this->dataPtr.get());
+
+  ignmsg << "JointController subscribing to Double messages on [" << topic
+         << "]" << std::endl;
+}
+
+//////////////////////////////////////////////////
+void JointController::PreUpdate(const ignition::gazebo::UpdateInfo &_info,
+    ignition::gazebo::EntityComponentManager &_ecm)
+{
+  // If the joint hasn't been identified yet, look for it
+  if (this->dataPtr->jointEntity == kNullEntity)
+  {
+    this->dataPtr->jointEntity =
+        this->dataPtr->model.JointByName(_ecm, this->dataPtr->jointName);
+  }
+
+  if (this->dataPtr->jointEntity == kNullEntity)
+    return;
+
+  // Nothing left to do if paused.
+  if (_info.paused)
+    return;
+
+  std::lock_guard<std::mutex> lock(this->dataPtr->jointVelCmdMutex);
+
+  // Update joint velocity
+  auto vel =
+      _ecm.Component<components::JointVelocity>(this->dataPtr->jointEntity);
+
+  if (vel == nullptr)
+  {
+    _ecm.CreateComponent(this->dataPtr->jointEntity,
+                         components::JointVelocity(this->dataPtr->jointVelCmd));
+  }
+  else
+  {
+    vel->Data() = this->dataPtr->jointVelCmd;
+  }
+}
+
+//////////////////////////////////////////////////
+void JointControllerPrivate::OnCmdVel(const msgs::Double &_msg)
+{
+  std::lock_guard<std::mutex> lock(this->jointVelCmdMutex);
+  this->jointVelCmd = _msg.data();
+}
+
+IGNITION_ADD_PLUGIN(JointController,
+                    ignition::gazebo::System,
+                    JointController::ISystemConfigure,
+                    JointController::ISystemPreUpdate)
+
+IGNITION_ADD_PLUGIN_ALIAS(JointController,
+                          "ignition::gazebo::systems::JointController")