+++ test/plugins/TestModelSystem.hh
@@ -31,9 +31,10 @@ namespace gazebo
 inline namespace IGNITION_GAZEBO_VERSION_NAMESPACE {
 namespace components
 {
-using IntComponent = components::Component<int, class IntComponentTag>;
-IGN_GAZEBO_REGISTER_COMPONENT("ign_gazebo_components.IntComponent",
-    IntComponent)
+using ModelPluginComponent = components::Component<int,
+    class ModelPluginComponentTag>;
+IGN_GAZEBO_REGISTER_COMPONENT("ModelPluginComponent",
+    ModelPluginComponent)
 }
 }
 
@@ -41,7 +42,15 @@ class TestModelSystem :
   public System,
   public ISystemConfigure
 {
-  public: TestModelSystem() = default;
+  public: TestModelSystem()
+        {
+          igndbg << "Constructing TestModelSystem" << std::endl;
+        }
+
+  public: ~TestModelSystem()
+        {
+          igndbg << "Destroying TestModelSystem" << std::endl;
+        }
 
   private: bool Service(msgs::StringMsg &_msg)
            {
@@ -54,13 +63,18 @@ class TestModelSystem :
                          EntityComponentManager &_ecm,
                          EventManager &/*_eventManager*/) override
         {
+          igndbg << "Configuring TestModelSystem" << std::endl;
           this->model = Model(_entity);
 
           auto link = this->model.LinkByName(_ecm, "link_1");
-          // This plugin might have been attached to the box model in
+          // This plugin might have been attached to the models in
           // test/world/shapes.world.
           if (link == kNullEntity)
             link = this->model.LinkByName(_ecm, "box_link");
+          if (link == kNullEntity)
+            link = this->model.LinkByName(_ecm, "sphere_link");
+          if (link == kNullEntity)
+            link = this->model.LinkByName(_ecm, "cylinder_link");
 
           // Fail to create component if link is not found
           if (link == kNullEntity)
@@ -74,7 +88,8 @@ class TestModelSystem :
               &TestModelSystem::Service, this);
 
           auto value = _sdf->Get<int>("model_key");
-          _ecm.CreateComponent(_entity, components::IntComponent(value));
+          _ecm.CreateComponent(_entity,
+              components::ModelPluginComponent(value));
         }
 
   private: Model model;