+++ Changelog.md
@@ -2,6 +2,12 @@
 
 ### Ignition Gazebo 1.X.X
 
+1. More ign-msgs <-> SDF conversions: Inertial, Geometry, Material
+    * [Pull Request 251](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/251)
+
+1. Add a basic JointController system
+    * [Pull Request 246](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/246)
+
 1. Added command line options to configure distributed simulation. These
    will replace the environment variables.
     * [Pull Request 238](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/238)